Build: shrink preview comment
Implement what we discussed in
https://github.com/readthedocs/readthedocs.org/pull/12295.
This is not perfect and it doesn't pretend to be, it's just one of the
implementations after a few discussion iterations in the comments.

The idea of this PR is to have something a little better than what we have to be
able to merge and move forward with.