Build: Denormalize latest build for projects (#12392)
Sorting by the last build project is expensive, as the query is done for
each project (even if paginated, since the order applies to all
projects). For comparison, here are some numbers with data from
production

```
# Ordering by name
145 ms ± 446 μs per loop (mean ± std. dev. of 5 runs, 5 loops each)

# Ordering by max(builds_date)
4.37 s ± 254 ms per loop (mean ± std. dev. of 5 runs, 5 loops each)
```

The migration is going to take some time to run, I imagine... but should
be worth it for a snappier dashboard.