update readthedocs documentation to ubuntu 24.04 (#12401)
I came from
https://docs.readthedocs.com/platform/stable/config-file/v2.html and
found that the recommended OS is outdated.
I think that this should always use an up-to-date version, so that new
users do not start with an old OS version for their docs.

Of course there are many other links to old OS versions, some also to
Ubuntu 20.04, which should be changed as well.