Builds: support custom Git checkout command
Initial POC to support this feature. It will allow us to support customers with
specific needs. We don't plan to expose this directly to users yet, since there
are some internals that need to be met to work:

- Use specific environment variables
- Clone the repository in `.` to keep with the default workflow
- Depending on the repository structure, a custom YAML path and custom
  `python.install.requirements` may be needed as well.

All of this can be documented eventually, but we need an internal way to write
these custom commands to onboard customers with these needs.

Closes #12313