Builds: support custom Git checkout command (#12412)
Initial POC to support this feature. It will allow us to support
customers with specific needs. We don't plan to expose this directly to
users yet, since there are some internals that need to be met to work:

- Use specific environment variables
- Clone the repository in `.` to keep with the default workflow
- Depending on the repository structure, a custom YAML path and custom
`python.install.requirements` may be needed as well (I added this commit
https://github.com/humitos/rocm-libraries/commit/a3938120dc6ef77bda31851fd5ad9bc176552f5f)

All of this can be documented eventually, but we need an internal way to
write these custom commands to onboard customers with these needs.

<img width="1161" height="518" alt="Screenshot_2025-08-15_12-06-48"
src="https://github.com/user-attachments/assets/b95c5007-5346-4539-883b-58546ee15d4f"
/>

The value for `Project.git_checkout_command` is:

```json
[
    "env",
    "echo $READTHEDOCS_GIT_CLONE_URL",
    "git clone --no-checkout --no-tag --filter=blob:none --depth 1 $READTHEDOCS_GIT_CLONE_URL .",
    "git sparse-checkout init --cone",
    "git sparse-checkout set projects/rocblas",
    "git checkout $READTHEDOCS_GIT_IDENTIFIER"
]
```

Closes #12313