Builds: healthcheck permissions (#12423)
We want to skip the default permsissions check here (done at
`get_queryset()` method) because it filters the object by build API key
and/or user access.

Since we are making an anonymous request without a build API key, we
need to skip this check and base it only on the `?builder=` GET
attribute.