Git services: allow updating a single repository (#12429)
When syncing permissions for SSO, we don't need to sync all repositories
the user has access to, we just need to update the permissions of the
repositories that are connected to projects that belong to SSO
organizations. So this adds a way to just update one single repository
at a time for all providers. We don't need to do this for GH app
projects, since they are kept up to date via a webhook.

Since this task should be faster now, we could run it more frequently,
so permissions are reflected faster on RTD.