Projects: improve deletion (#12456)
For some related fields, Django first does a select, then it probably
loads each object into memory? But after that it does the deletion using
a `IN` statement with the ID of the objects...

Using raw_delete avoids that, but it can also cause integrity problems,
as Django does its own implementation of on_delete, so I chose models
that don't have additional relations and also shouldn't have problems as
they are always deleted when the project is deleted. Another source of
problems is when models have foreign fields with on_delete=SET_NULL, as
django will do an UPDATE over the model with all the related fields...
for example for latest_build, django will run an update over all
projects with an `IN (... all builds IDs..)`, to set the latest_build ID
to null..., which is silly since the whole project model is being
deleted, but django doesn't know that lates_build can only contain a
build from the project. A raw delete could help, but builds and versions
have a lot of related fields in other models, so I'm a little hesitant.

Closes https://github.com/readthedocs/readthedocs.org/pull/12410
Ref https://github.com/readthedocs/readthedocs.org/issues/10040