Project: use latest version for syncing versions (#12461)
We should refactor this task so it doesn't require a version, but that's
a bigger change... This is basically the same as it was before, as
LATEST is the only version that its identifier is the same as the
default branch.

I'm doing small refactors to rely less on get_default_branch, as that
contains some extra logic that we already run to have latest always in
sync.