Build: don't reset `healthcheck` when updating builds
When the build process finishes, we update the `Build` object via the API with
new values (e.g. length) re-using the `APIBuild` object we fetched via the API
before starting the build.

Since that object contains `APIBuild.healthcheck` _we are resetting_ the
healthcheck before finishing the build.

This commit removes the `healthcheck` field from the `APIBuild` returned, so
it's not updated by the build while it's being executed.