Builds: save `task_executed_at` to help debugging (#12500)
Save the datetime when the build was executed by a worker and log a
warning if it's bigger than 10 minutes (we can adjust this number as we
want). Note that 5 minutes is the default retry time.

This is a POC to help debugging
https://github.com/readthedocs/readthedocs.org/issues/12472

This data together with other logs like
* "Delaying tasks due to concurrency limit."
* "Retrying this task."

could help us to understand what's going on here.