Version sync: skip syncing versions when lsremote fails (#12540)
Currently, if lsremote fails, we are returning an empty list of branches
and tags, which is interpreted by our system as if the branches/tags
were deleted, which is incorrect.

I was thinking on raising this exception to the user, but I decided
against it, as the user won't be able to fix the problem (it could be a
temporal issue), if the repo isn't correctly configured, they will get
the error when trying to clone the repo.

---------

Co-authored-by: Manuel Kaufmann <humitos@gmail.com>