Pass ProjectVersionInfo as JSON-safe data
In order to safely pass the data across the celery task `defer` call,
which implicitly JSON de/serializes data, the objects need to be
converted to and from a JSON-compatible shape.

As a simple paradigm for handling this, `serialize_many` and `parse_many`
are added to `ProjectVersionInfo` as classmethods for converting to and
from a list of dicts.
By structuring `ProjectVersionInfo` as a decorator, it's possible to
keep the code within the task definition the same as it was before.