Build: better support for empty default branch (#12462)
The basic idea is to have the value of latest in sync with the value
from the default branch, which we already do. But, in case the default
branch is left empty, we were falling back to master, which is just a
guess of the default, so instead of guessing the value, we just don't
update it if we don't know its value. When cloning the repo, we have
access to the default branch, so that's when we update latest for those
projects.

Note this is only relevant for projects that aren't linked to a remote
repository, since for those we know the default branch. So, when we try
to build the latest version of a project that doesn't have a default
branch, we:

- Omit the ref from the fetch command.
- Use the default branch from the repository for the checkout (we could
skip this step, but since we have a new feature that keeps the files
from a previous clone, we need to do a checkout to make sure we are
building from the given branch/commit).
- Update the identifier from latest to the default branch we detected
from the cloned repository.

There's still the problem about latest's identifier being out of sync
when the default branch changes, but that's fixed after a new build, and
this isn't a new problem (now users aren't blocked at least), and
changing the default branch of a repo isn't something users do that
often, and this only affects projects that aren't linked to a
repository, so I think we are okay with that for now.

This replaces https://github.com/readthedocs/readthedocs.org/pull/10927/
Closes https://github.com/readthedocs/readthedocs.org/issues/12231