Search: record queries in batch (#12660)
When doing a search over several projects (like subprojects), we are
creating a task for each project. If a project has several subprojects,
this will create a task for each one, instead of doing that, we accept
several projects in the task.

This will reduce the number of calls to redis, which adds up to the
response time.
<img width="1712" height="403" alt="Screenshot 2025-12-17 at 16-59-41
Read the Docs com Read the Docs com New Relic"
src="https://github.com/user-attachments/assets/dcc01346-fd92-4eb7-a6b2-0a2f7dbec18c"
/>